pyproject.toml
[project]
version = "0.1.0"
name = "dataset_foundry"
description = "Build and refine datasets using LLMs"
readme = "README.md"
authors = [
  {name = "Trevor Lohrbeer", email = "trevor@fastfedora.com"},
]
requires-python = ">=3.12"
dependencies = [
    "langchain==0.3.17",
    "langchain-anthropic==0.3.6",
    "langchain-core==0.3.33",
    "langchain-openai==0.3.3",
    "pydantic>=2.10.6",
    "pytest>=8.3.4",
    "python-dotenv==1.0.1",
    "PyYAML==6.0.2",
    "toolz>=1.0.0",
]

[project.scripts]
dataset-foundry = "src.dataset_foundry.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
